REFACTOR test assets to 'tests/Assets' directory
