+++ tests/Assets/Models/CompanyPeople.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Sfneal\Models\Tests\Models;
+namespace Sfneal\Models\Tests\Assets\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Sfneal\Models\Model;
-use Sfneal\Models\Tests\Factories\CompanyPeopleFactory;
+use Sfneal\Models\Tests\Assets\Factories\CompanyPeopleFactory;
 
 class CompanyPeople extends Model
 {
