+++ tests/TestCase.php
@@ -5,7 +5,8 @@ namespace Sfneal\Models\Tests;
 use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
-use Sfneal\Models\Tests\Providers\TestingServiceProvider;
+use Sfneal\Models\Tests\Assets\Providers\TestingServiceProvider;
+
 
 class TestCase extends OrchestraTestCase
 {
@@ -32,7 +33,7 @@ class TestCase extends OrchestraTestCase
      */
     protected function getEnvironmentSetUp($app)
     {
-        include_once __DIR__.'/migrations/create_people_table.php.stub';
+        include_once __DIR__ . '/Assets/migrations/create_people_table.php.stub';
 
         (new \CreatePeopleTable())->up();
     }
